rag-service/requirements.txt
fastapi
uvicorn
requests
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
